@@ -1,11 +1,12 @@
 using Api.Data;
-using DataObjects;
+using SharedLibrary.Entities;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
 
 namespace Api.Controllers;
 
-[Route("api/[controller]")]
 [ApiController]
+[Route("api/[controller]")]
 public class CottageController : ControllerBase
 {
     private readonly ApplicationDbContext _context;
@@ -16,9 +17,9 @@ public class CottageController : ControllerBase
     }
 
     [HttpGet("all")]
-    public ActionResult GetCottages()
+    public async Task<ActionResult<List<Cottage>>> GetCottages()
     {
-        var cottages = _context.Cottages.ToList();
+        var cottages = await _context.Cottages.ToListAsync();
 
         if (cottages is null || cottages.Count == 0)
         {
@@ -29,51 +30,81 @@ public class CottageController : ControllerBase
     }
 
     [HttpGet("{id:int}")]
-    public async Task<IActionResult> GetCottageAsync(int id)
+    public async Task<ActionResult<Cottage>> GetCottage(int id)
     {
         var cottage = await _context.Cottages.FindAsync(id);
+        
+        if (cottage is null)
+        {
+            return NotFound(id);
+        }
+        
         return Ok(cottage);
     }
 
-    [HttpPut]
-    public async Task<IActionResult> UpdateCottageAsync(int id, Cottage cottage)
+    [HttpPut("")]
+    public async Task<ActionResult<List<Cottage>>> UpdateCottage(Cottage cottage)
     {
-        var c = _context.Cottages.Find(id);
-        if (c == null)
+        var cottageUpdate = await _context.Cottages.FindAsync(cottage.Id);
+        if (cottageUpdate == null)
         {
-            return BadRequest();
+            return NotFound();
         }
 
-        _context.Cottages.Update(cottage);
+        cottageUpdate.Title = cottage.Title;
+        cottageUpdate.Description = cottage.Description;
+        cottageUpdate.CostPerDay = cottage.CostPerDay;
+        cottageUpdate.PictureUrl = cottage.PictureUrl;
+        cottageUpdate.Rooms = cottage.Rooms;
+        cottageUpdate.IsElectricity = cottage.IsElectricity;
+        cottageUpdate.IsSauna = cottage.IsSauna;
+
         await _context.SaveChangesAsync();
-        return Ok();
+        return Ok(await _context.Cottages.ToListAsync());
     }
 
     [HttpPost]
-    public async Task CreateCottageAsync(Cottage input)
+    public async Task<ActionResult<List<Cottage>>> CreateCottage(Cottage newCottage)
     {
-        var cottage = new Cottage
-        {
-            Title = input.Title,
-            Description = input.Description,
-            CostPerDay = input.CostPerDay
-        };
-
-        await _context.Cottages.AddAsync(cottage);
+        _context.Cottages.Add(newCottage);
         await _context.SaveChangesAsync();
+
+        return Ok(await _context.Cottages.ToListAsync());
     }
 
     [HttpDelete("{id:int}")]
-    public async Task<IActionResult> DeleteCottageAsync(int id)
+    public async Task<ActionResult<int>> DeleteCottage(int id)
     {
         var cottage = await _context.Cottages.FindAsync(id);
-        if (cottage is null)
+        if (cottage == null)
         {
             return NotFound(id);
         }
         
         _context.Cottages.Remove(cottage);
         await _context.SaveChangesAsync();
-        return Ok(cottage);
+        return Ok();
+    }
+
+    [HttpGet("{id:int}/Reservations")]
+    public async Task<ActionResult<Reservation>> GetReservations(int id)
+    {
+        var cottage = await _context.Cottages.FirstOrDefaultAsync(c => c.Id == id);
+        
+        if (cottage is null)
+        {
+            return NotFound(id);
+        }
+        
+        var reservations = cottage.Reservations;
+
+        if (reservations.Count > 0 || reservations is not null)
+        {
+            return Ok(reservations);
+        }
+        else 
+        {
+            return NotFound(id);
+        }
     }
 }
