@@ -23,11 +23,11 @@ namespace Backend
         {
             services.AddRazorPages();
             services.AddServerSideBlazor();
-            services.AddSingleton<CottageService>();
             services.AddHttpClient<ICottageService, CottageService>(client =>
             {
-                client.BaseAddress = new Uri("http://localhost:5260/");
+                client.BaseAddress = new Uri(Configuration.GetValue<string>("ApiBaseAddress"));
             });
+            services.AddScoped<CottageService>();
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
